6_django/requirements.txt
Django==4.0.5
psycopg2-binary==2.9.3
djangorestframework==3.14.0